Renamed globals.py to utils.py
